messages: Clean up and align with Centrifuge chain
We merge all the test*Encoding and test*Decoding tests in one test* test
where we essentially test the identity property of the format & parse
functions for a given message. This reduces the number of test cases and
the amount of duplicated data across them.

Here, we also align the tests with the Centrifuge Chain by using the
values used over there and therefore verify that both systems, given the
same message, encode them to the exact same hex and decode that same hex
back to the original message.